+++ Examples/ObjectiveC/GDPRExample/GDPRExample/UserDefaultsKeys.m
@@ -1,0 +1,11 @@
+/*
+ * Version for iOS © 2015–2018 YANDEX
+ *
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://yandex.com/legal/mobileads_sdk_agreement/
+ */
+
+#import "UserDefaultsKeys.h"
+
+NSString *const kGDPRUserConsentKey = @"userConsent";
+NSString *const kGDPRShouldShowDialogKey = @"shouldShowDialog";